+++ cms/djangoapps/contentstore/api/__init__.py
@@ -1,0 +1,1 @@
+from .views.utils import course_author_access_required
