+++ cms/djangoapps/contentstore/api/urls.py
@@ -6,7 +6,9 @@ from django.urls import re_path
 
 from cms.djangoapps.contentstore.api.views import course_import, course_quality, course_validation
 
+
 app_name = 'contentstore'
+helper = "{0,1}"
 
 urlpatterns = [
     re_path(fr'^v0/import/{settings.COURSE_ID_PATTERN}/$',
@@ -15,4 +17,5 @@ urlpatterns = [
             course_validation.CourseValidationView.as_view(), name='course_validation'),
     re_path(fr'^v1/quality/{settings.COURSE_ID_PATTERN}/$',
             course_quality.CourseQualityView.as_view(), name='course_quality'),
+
 ]
