+++ cms/djangoapps/contentstore/rest_api/v1/views/__init__.py
@@ -1,0 +1,2 @@
+from .proctored_exam_settings import ProctoredExamSettingsView
+from .xblock import XblockView
