+++ cms/djangoapps/contentstore/views/tests/test_block.py
@@ -49,7 +49,7 @@ from xmodule.x_module import STUDENT_VIEW, STUDIO_VIEW
 
 from cms.djangoapps.contentstore.tests.utils import CourseTestCase
 from cms.djangoapps.contentstore.utils import reverse_course_url, reverse_usage_url
-from cms.djangoapps.contentstore.views import block as item_module
+from cms.djangoapps.contentstore.xblock_services import xblock_service as item_module
 from common.djangoapps.student.tests.factories import StaffFactory, UserFactory
 from common.djangoapps.xblock_django.models import (
     XBlockConfiguration,
@@ -61,10 +61,10 @@ from lms.djangoapps.lms_xblock.mixin import NONSENSICAL_ACCESS_RESTRICTION
 from openedx.core.djangoapps.discussions.models import DiscussionsConfiguration
 
 from ..component import component_handler, get_component_templates
-from ..block import (
+from cms.djangoapps.contentstore.xblock_services.xblock_service import (
     ALWAYS,
     VisibilityState,
-    _get_block_info,
+    get_block_info,
     _get_source_index,
     _xblock_type_and_display_name,
     add_container_page_publishing_info,
@@ -90,6 +90,7 @@ class AsideTest(XBlockAside):
 
 class ItemTest(CourseTestCase):
     """ Base test class for create, save, and delete """
+
     def setUp(self):
         super().setUp()
 
@@ -451,7 +452,7 @@ class GetItemTest(ItemTest):
                     xblock = parent_xblock
             else:
                 self.assertNotIn('ancestors', response)
-                self.assertEqual(_get_block_info(xblock), response)
+                self.assertEqual(get_block_info(xblock), response)
 
 
 @ddt.ddt
@@ -546,6 +547,7 @@ class DuplicateHelper:
     """
     Helper mixin class for TestDuplicateItem and TestDuplicateItemWithAsides
     """
+
     def _duplicate_and_verify(self, source_usage_key, parent_usage_key, check_asides=False):
         """ Duplicates the source, parenting to supplied parent. Then does equality check. """
         usage_key = self._duplicate_item(parent_usage_key, source_usage_key)
@@ -1248,7 +1250,7 @@ class TestMoveItem(ItemTest):
         validation = html.validate()
         self.assertEqual(len(validation.messages), 0)
 
-    @patch('cms.djangoapps.contentstore.views.block.log')
+    @patch('cms.djangoapps.contentstore.xblock_services.xblock_service.log')
     def test_move_logging(self, mock_logger):
         """
         Test logging when an item is successfully moved.
@@ -1909,6 +1911,7 @@ class TestEditItemSplitMongo(TestEditItemSetup):
     """
     Tests for EditItem running on top of the SplitMongoModuleStore.
     """
+
     def test_editing_view_wrappers(self):
         """
         Verify that the editing view only generates a single wrapper, no matter how many times it's loaded
@@ -2579,6 +2582,7 @@ class TestXBlockInfo(ItemTest):
     """
     Unit tests for XBlock's outline handling.
     """
+
     def setUp(self):
         super().setUp()
         user_id = self.user.id
